requirements.txt
# Flask Web Framework
Flask==2.0.3
Werkzeug==2.0.3

# Core dependencies
torch>=2.0.0
transformers>=4.30.0
opencv-python>=4.8.1.78
Pillow>=10.0.1
numpy>=1.21.0

# Optional: For better performance
# gunicorn==21.2.0  # For production deployment

# Development dependencies (optional)
# python-dotenv==1.0.0  # For environment variables
